pyproject.toml
[project]
name = "leafmine-segmentation"
version = "0.1.0"
description = "U-Net segmentation pipeline for leafmine detection"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "albumentations>=1.4.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0.0",
    "scikit-learn>=1.3.0",
    "segmentation-models-pytorch>=0.3.3",
    "tensorboard>=2.14.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "tqdm>=4.66.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/leafmine_seg"]
